Split DefaultVersionStrategy in two
