+++ src/WebApiContrib.Core.Versioning/Strategies/AcceptHeaderParameterVersionStrategy.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Linq;
 using Microsoft.Net.Http.Headers;
 
 namespace WebApiContrib.Core.Versioning
@@ -14,10 +13,8 @@ namespace WebApiContrib.Core.Versioning
     /// 2. Using a facet-based sub-type with the version as the last facet; <c>application/vnd.my-app.v2</c>
     /// </para>
     /// </summary>
-    public class DefaultVersionStrategy : AcceptHeaderVersionStrategy
+    public class AcceptHeaderParameterVersionStrategy : AcceptHeaderVersionStrategy
     {
-        private static readonly char[] DotSeparator = {'.'};
-
         public string ParameterName { get; set; } = "version";
 
         /// <inheritdoc />
@@ -36,30 +33,12 @@ namespace WebApiContrib.Core.Versioning
             }
 
             int? version;
-
             if (TryGetParameterVersion(acceptHeader, ParameterName, out version))
             {
                 return version;
             }
 
-            if (TryGetFacetVersion(subType, out version))
-            {
-                return version;
-            }
-
             return null;
-        }
-
-        private static string StripSuffix(string subType)
-        {
-            var suffixSeparatorIndex = subType.IndexOf('+');
-
-            if (suffixSeparatorIndex >= 0)
-            {
-                return subType.Substring(0, suffixSeparatorIndex);
-            }
-
-            return subType;
         }
 
         private static bool TryGetParameterVersion(MediaTypeHeaderValue acceptHeader, string parameterName, out int? version)
@@ -69,27 +48,6 @@ namespace WebApiContrib.Core.Versioning
                 if (parameter.Name.Equals(parameterName, StringComparison.OrdinalIgnoreCase))
                 {
                     if (ParsingUtility.TryParseVersion(parameter.Value, out version))
-                    {
-                        return true;
-                    }
-                }
-            }
-
-            version = null;
-            return false;
-        }
-
-        private static bool TryGetFacetVersion(string subType, out int? version)
-        {
-            var facetSeparatorIndex = subType.IndexOf('.');
-
-            if (facetSeparatorIndex >= 0)
-            {
-                var facets = subType.Split(DotSeparator).Reverse();
-
-                foreach (var facet in facets)
-                {
-                    if (ParsingUtility.TryParseVersion(facet, out version))
                     {
                         return true;
                     }
